Audio Fixes
DestroyOnLoad Saved my ass and now we can have a single AudioManager GameObject for both scenes and bg music doesn't stop when switching from one scene to another.
Made AudioData & added it to the SavingSystem
Fixed the volume sliders value to load according to the correct (saved) values on startup.